feat: add Prune trait
This is going to be used to prune the DAGs. For now it's not very
useful. But, when we have imp, specialisation, we can incorporate this
at every level of the Compile process. This will result in highly
compressed DAGs which will hopefully, get around the type explosion.

Eventually we may want to add another trait, say `Simplify`, which uses
`Prune` to reduce the expressions further still.