feat: rename State -> Database
While the name "State" was nice, it had some issues. First of all, this
crate will likely be used in `rsrl` which has a distinct notion of
state. This would have caused confusion. Furthermore, state implies that
things are changing, which they may not be in this context.

As a result, it made more sense to replace State with Database since,
according to the dictionary definition, is the most appropriate word.

This commit also includes some general improves and additional features.