feat(buffers): add Buffer::mutate method
Buffer::mutate is the in-place variant of map/map_ref. It applies a
function to each element of the buffer, in-place, so as not to allocate
any new memory if not needed.

An important distinction here is that mutate __cannot__ modify the
Buffer::Field type. But, it is also much more efficient when no change
is made to the scalar field.