feat: optimisations for Spec::Diagonal in linalg
This commit includes some new spec optimisation rules. Namely, for
linear algebraic contractions, we now take into account Spec::Full,
especially if the value is zero, and Spec::Diagonal. The latter is nice
because multiplications of the form Ab or bA, where A \in (N, M) or (M,
N) and b \in (N), can be dramatically simpler if we know A is diagonal.

This commit leads to a ~40% improvement in performance for the lr_sgd
benchmark.